3.9.23

+++ src/org/hy/common/xml/plugins/XSQLGroup.java
@@ -3001,7 +3001,7 @@ public final class XSQLGroup
 
 
         @Override
-        public int getSerialNo()
+        public long getSerialNo()
         {
             return GetSerialNo();
         }
